backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
python-multipart
pydantic-settings
sqlalchemy[asyncio]
aiosqlite
groq
elevenlabs
